+++ tests/Unit/Css/CssDocumentTest.php
@@ -529,7 +529,6 @@ final class CssDocumentTest extends TestCase
         $result = $subject->renderNonConditionalAtRules();
 
         preg_match('/@[\\w\\-]++/', $atRuleCss, $atAndRuleNameMatches);
-        self::assertIsArray($atAndRuleNameMatches);
         $atAndRuleName = $atAndRuleNameMatches[0];
         self::assertStringNotContainsString($atAndRuleName, $result);
     }