+++ CHANGELOG.md
@@ -10,6 +10,8 @@ Please also have a look at our
 
 ### Added
 
+- Add support for Symfony 8.0 (#1509)
+
 ### Changed
 
 - Improve performance of checking for existing `Content-Type` (#1518, #1519)